+++ bot/common/commands.py
@@ -1,21 +1,47 @@
+import json
 import logging
 import discord
 import hashlib
+import constants
 
-from airtable import find_user, create_user, get_discord_record, get_guild
-from cache import get_thread, build_cache_value, StepKeys, ThreadKeys, Onboarding
+intents = discord.Intents.all()
+bot = discord.Bot(intents=intents)
+
+
+from common.core import bot
+from common.airtable import find_user, create_user, get_discord_record, get_guild
+from common.threads.thread_builder import (
+    build_cache_value,
+    ThreadKeys,
+)
+from common.threads.onboarding import Onboarding
+from common.threads.update import UpdateProfile
 from config import read_file, GUILD_IDS, INFO_EMBED_COLOR, Redis, get_list_of_emojis
 from exceptions import NotGuildException
 from exceptions import ErrorHandler
 
 logger = logging.getLogger(__name__)
 
-intents = discord.Intents.all()
-bot = discord.Bot(intents=intents)
+
+def is_enabled():
+    return constants.bot.is_dev
+
+
+def get_thread(user_id, key):
+    val = json.loads(key)
+    thread = val.get("thread")
+    step = val.get("step")
+    message_id = val.get("message_id")
+    guild_id = val.get("guild_id")
+    if thread == ThreadKeys.ONBOARDING.value:
+        return Onboarding(user_id, step, message_id, guild_id)
+    elif thread == ThreadKeys.UPDATE_PROFILE.value:
+        return UpdateProfile(user_id, step, message_id, guild_id)
+    raise Exception("Unknown Thread!")
 
 
 @bot.slash_command(
-    guild_id=GUILD_IDS, description="Send users link to report engagement"
+    guild_id=GUILD_IDS, description="Send users link to report engagement",
 )
 async def report(ctx):
     is_guild = bool(ctx.guild)
@@ -39,24 +65,6 @@ async def report(ctx):
         ctx.response.is_done()
 
 
-# Join should be pretty simple
-# If the user exists in airtable
-# then send a message asking about their adventure
-# If they aren't a user dm
-# In DM send a prompt and start asking for information
-# If a user sends a random message then tell them to continue
-# on the flow, if they repond with an unsupported emjoi
-# do the same.
-#
-# Keep an internal dictionary to keep track of what thread they are on
-
-
-# If not a user
-# Check if bot can DM
-# If false than tell user to enable dms
-# If true than DM requesting information
-
-
 @bot.slash_command(guild_id=GUILD_IDS, description="Get started with Govern")
 async def join(ctx):
     is_guild = bool(ctx.guild)
@@ -64,16 +72,13 @@ async def join(ctx):
         raise NotGuildException("Command was executed outside of a guild")
 
     is_user = await find_user(ctx.author.id, ctx.guild.id)
-    print(is_user)
     if is_user:
         # Send welcome message and
         # And ask what journey they are
         # on by sending all the commands
         application_commands = bot.application_commands
         embed = discord.Embed(
-            colour=INFO_EMBED_COLOR,
-            title="Welcome Back",
-            description=f"",
+            colour=INFO_EMBED_COLOR, title="Welcome Back", description="",
         )
         for cmd in application_commands:
             if isinstance(cmd, discord.SlashCommand):
@@ -86,8 +91,6 @@ async def join(ctx):
 
     await ctx.response.defer()
     # store guild_id and disord_id
-    print(ctx.guild.id)
-    print(ctx.author.id)
     await create_user(ctx.author.id, ctx.guild.id)
     # check if user can be DMed
     can_send_message = ctx.can_send(discord.Message)
@@ -97,18 +100,14 @@ async def join(ctx):
         )
         return
 
-    # Get guild
-
-    # Send messsage explaining we need to get some
-    # Add a bunch of fields inline
     embed = discord.Embed(
         colour=INFO_EMBED_COLOR,
         title="Welcome",
-        description=f"Thank you for joining the Govrn ecosystem! To help automate  gathering your contributions to {ctx.guild.name} we need you to provide some information. Any of the following data requests can be skipped with the ⏭️  emoji!",
+        description="Thank you for joining the Govrn ecosystem! "
+        "To help automate  gathering your contributions to {ctx.guild.name} "
+        "we need you to provide some information. Any of the following data "
+        "requests can be skipped with the ⏭️  emoji!",
     )
-    # Add user to the internal cache with with appropriate stage type
-    # Thread Type, thread type will have a numer of steps,
-    # key:user -> value: thread_type+step
     logger.info(
         f"Key: {build_cache_value(ThreadKeys.ONBOARDING.value, '', ctx.guild.id)}"
     )
@@ -123,13 +122,6 @@ async def join(ctx):
     guild_id=GUILD_IDS, description="Update your profile for a given Dao"
 )
 async def update(ctx):
-    # If sent from guild see if user has onboarded
-    # If user has onboarded show them their profile and tell them how to update with reactions
-    #
-    # If DM show user a bunch of guilds they are in and they can choose which to update
-    # by selecting an emoji
-    #
-    # After updating a field send a thank you
     is_guild = bool(ctx.guild)
     if is_guild:
         return
@@ -139,7 +131,9 @@ async def update(ctx):
         if not airtable_guild_ids:
             embed = discord.Embed(
                 colour=INFO_EMBED_COLOR,
-                description=f"Cannot update profile because you have not been onboarded to any daos. Run /join in the discord you want to onboard to",
+                description="Cannot update profile because you "
+                "have not been onboarded to any daos. Run /join in the "
+                "discord you want to onboard to",
             )
             await ctx.response.send_message(embed=embed)
             ctx.response.is_done()
@@ -149,14 +143,13 @@ async def update(ctx):
         guild_ids = []
         for record_id in airtable_guild_ids:
             g = await get_guild(record_id)
-            print(g)
             guild_id = g.get("guild_id")
             if guild_id:
                 guild_ids.append(guild_id)
         embed = discord.Embed(
             colour=INFO_EMBED_COLOR,
             title="Welcome",
-            description=f"Which profile guild would you like to update?",
+            description="Which profile guild would you like to update?",
         )
         emojis = get_list_of_emojis(len(guild_ids))
         daos = {}
@@ -170,8 +163,6 @@ async def update(ctx):
         message = await ctx.followup.send(embed=embed)
         for emoji in emojis:
             await message.add_reaction(emoji)
-        print("Author")
-        print(ctx.author)
         await Redis.set(
             ctx.author.id,
             build_cache_value(
@@ -187,8 +178,6 @@ async def update(ctx):
 # Event listners
 @bot.event
 async def on_application_command_error(ctx, exception):
-    # All commands will be slash commands
-    # Commands will be sent back from where they come from
     err = ErrorHandler(exception)
     logger.info(f"Command error type {type(exception)}")
     await ctx.response.send_message(err.msg)
@@ -197,16 +186,13 @@ async def on_application_command_error(ctx, exception):
 
 @bot.event
 async def on_message(message):
-    print("Message")
     if message.author.bot is True:
         return
 
-    print("Is DM")
     # Check channel DM channel
     if not isinstance(message.channel, discord.DMChannel):
         return
 
-    print("Getting thread key")
     # Check if user has open thread
     thread_key = await Redis.get(message.author.id)
     if not thread_key:
@@ -214,11 +200,7 @@ async def on_message(message):
         return
 
     thread = get_thread(message.author.id, thread_key)
-    print(thread)
     await thread.send(message)
-
-    print(message)
-    print(message.author)
 
 
 @bot.event
@@ -231,29 +213,18 @@ async def on_raw_reaction_add(payload):
     if user.bot is True:
         return
 
-    print("first line")
-    print(payload)
-    print(reaction)
-    print(user)
-    print(channel)
-    print(reaction.channel_id)
     # Check channel DM channel
     if not isinstance(channel, discord.DMChannel):
         return
 
     # Check if user has open thread
     thread_key = await Redis.get(user.id)
-    print("ThreadKey")
-    print(thread_key)
     if not thread_key:
         # TODO: It may make sense to send some sort of message here
         return
 
     thread = get_thread(user.id, thread_key)
     await thread.handle_reaction(reaction, user)
-
-    print(reaction)
-    print("hi")
 
 
 bot.on_application_command_error = on_application_command_error
