+++ bot/common/core.py
@@ -1,0 +1,6 @@
+import discord
+
+intents = discord.Intents.all()
+bot = discord.Bot(intents=intents)
+
+import common.commands
